src/requirements.txt
faust-streaming==0.10.15
python-schema-registry-client==2.5.0
python-dateutil==2.8.2